# fix(pointcloud_preprocessor): handle empty pointclouds in pickup_based_downsample_filter