## Description

The `pickup_based_downsample_filter` currently fails to publish an output when it receives an empty pointcloud.

The expected behavior is to publish an empty pointcloud when the input is also empty. This PR fixes the node to ensure this behavior.

To verify the fix, I've added integration tests for both single-point and empty pointclouds.

**Note:** These integration tests are time-consuming. I plan to replace them with faster unit tests for the `filter()` function in a future PR.


## Related links

**Private Links:**

- [TIER IV internal link](https://tier4.atlassian.net/browse/RT2-2536)

## How was this PR tested?

### Unit test
Before this change, the integration test for empty pointclouds would fail:
```
8: [==========] Running 2 tests from 1 test suite.
8: [----------] Global test environment set-up.
8: [----------] 2 tests from PickupBasedVoxelGridDownsampleFilterTest
8: [ RUN      ] PickupBasedVoxelGridDownsampleFilterTest.TestPointCloudPublishing
8: [       OK ] PickupBasedVoxelGridDownsampleFilterTest.TestPointCloudPublishing (316 ms)
8: [ RUN      ] PickupBasedVoxelGridDownsampleFilterTest.TestEmptyPointCloudPublishing
8: [ERROR] [1752720381.588830558] [PickupBasedVoxelGridDownsampleFilterComponent]: Cloud does not contain x, y, z fields. Cannot transform.
8: /home/takahisaishikawa/autoware/src/universe/autoware_universe/sensing/autoware_pointcloud_preprocessor/test/test_pickup_based_voxel_grid_downsample_filter_node.cpp:157: Failure
8: Value of: received_output_
8:   Actual: false
8: Expected: true
8: Expected to receive empty output point cloud
8: [  FAILED  ] PickupBasedVoxelGridDownsampleFilterTest.TestEmptyPointCloudPublishing (5320 ms)
8: [----------] 2 tests from PickupBasedVoxelGridDownsampleFilterTest (5636 ms total)
8: 
8: [----------] Global test environment tear-down
8: [==========] 2 tests from 1 test suite ran. (5636 ms total)
8: [  PASSED  ] 1 test.
8: [  FAILED  ] 1 test, listed below:
8: [  FAILED  ] PickupBasedVoxelGridDownsampleFilterTest.TestEmptyPointCloudPublishing
8: 
8:  1 FAILED TEST
```

After this change, all tests pass successfully:
```
8: [==========] Running 2 tests from 1 test suite.               
8: [----------] Global test environment set-up.
8: [----------] 2 tests from PickupBasedVoxelGridDownsampleFilterTest
8: [ RUN      ] PickupBasedVoxelGridDownsampleFilterTest.TestPointCloudPublishing
8: [       OK ] PickupBasedVoxelGridDownsampleFilterTest.TestPointCloudPublishing (319 ms)
8: [ RUN      ] PickupBasedVoxelGridDownsampleFilterTest.TestEmptyPointCloudPublishing
8: [       OK ] PickupBasedVoxelGridDownsampleFilterTest.TestEmptyPointCloudPublishing (312 ms)
8: [----------] 2 tests from PickupBasedVoxelGridDownsampleFilterTest (631 ms total)
8: 
8: [----------] Global test environment tear-down
8: [==========] 2 tests from 1 test suite ran. (631 ms total)
8: [  PASSED  ] 2 tests.
```

### Test with python script
I also manually verified the fix using a Python script. The script publishes an empty pointcloud, and I confirmed that the `pickup_based_voxel_grid_downsample_filter_node` correctly publishes an empty output in response.
```sh
# Terminal 1
ros2 run autoware_pointcloud_preprocessor pickup_based_voxel_grid_downsample_filter_node --ros-args -p voxel_size_x:=1.0 -p voxel_size_y:=1.0 -p voxel_size_z:=1.0 -r input:=pointcloud

# Terminal 2
python3 pointcloud_publisher.py 1 0

# Terminal 3
~$ ros2 topic echo /output
header:
  stamp:
    sec: 1752719349
    nanosec: 601025354
  frame_id: base_link
height: 1
width: 0
fields:
- name: x
  offset: 0
  datatype: 7
  count: 1
- name: y
  offset: 4
  datatype: 7
  count: 1
- name: z
  offset: 8
  datatype: 7
  count: 1
is_bigendian: false
point_step: 12
row_step: 0
data: []
is_dense: true
---
```

pointcloud_publisher.py is following script.
```py
import rclpy
from rclpy.node import Node
from sensor_msgs.msg import PointCloud2, PointField
from std_msgs.msg import Header
import struct
import argparse

class SinglePointCloudPublisher(Node):

    def __init__(self, publish_rate, num_points):
        super().__init__('single_pointcloud_publisher')

        if publish_rate <= 0:
            self.get_logger().error("Publish rate must be greater than 0.")
            raise ValueError("Publish rate must be greater than 0.")

        if num_points < 0:
            self.get_logger().error("Number of points must be greater than 0.")
            raise ValueError("Number of points must be greater than 0.")

        self.num_points = num_points
        self.publisher_ = self.create_publisher(PointCloud2, 'pointcloud', 10)
        self.timer_period = 1.0 / publish_rate # タイマー周期を計算
        self.timer = self.create_timer(self.timer_period, self.timer_callback)

        self.get_logger().info(f'Publishing PointCloud2 with {num_points} points at {publish_rate} Hz.')

    def timer_callback(self):
        cloud_msg = PointCloud2()

        # ヘッダーの設定
        cloud_msg.header = Header()
        cloud_msg.header.stamp = self.get_clock().now().to_msg()
        cloud_msg.header.frame_id = 'base_link'

        # フィールドの設定（x, y, z座標）
        cloud_msg.fields = [
            PointField(name='x', offset=0, datatype=PointField.FLOAT32, count=1),
            PointField(name='y', offset=4, datatype=PointField.FLOAT32, count=1),
            PointField(name='z', offset=8, datatype=PointField.FLOAT32, count=1)
        ]

        # 点群の設定
        cloud_msg.height = 1
        cloud_msg.width = self.num_points
        cloud_msg.is_bigendian = False
        cloud_msg.point_step = 12  # x, y, z それぞれ4バイト
        cloud_msg.row_step = 12 * self.num_points
        cloud_msg.is_dense = True

        # 指定された数の点のデータを生成
        point_data = b''
        for i in range(self.num_points):
            # 各点を少しずつずらして配置
            x = float(i) * 0.1
            y = 0.0
            z = 0.0
            point_data += struct.pack('<fff', x, y, z)
        
        cloud_msg.data = point_data

        self.publisher_.publish(cloud_msg)

def main(args=None):
    rclpy.init(args=args)

    # argparseを使ってコマンドライン引数を解析
    parser = argparse.ArgumentParser(description='Publish PointCloud2 messages')
    parser.add_argument('publish_rate', type=float, help='Publishing rate in Hz')
    parser.add_argument('num_points', type=int, help='Number of points in the point cloud')
    
    parsed_args = parser.parse_args()
    
    try:
        node = SinglePointCloudPublisher(parsed_args.publish_rate, parsed_args.num_points)
        rclpy.spin(node)
        node.destroy_node()
        rclpy.shutdown()
    except ValueError as e:
        print(f"Error: {e}")
        parser.print_help()

if __name__ == '__main__':
    main()
```

## Notes for reviewers

None.

## Interface changes

None.